Compiling after applying stash
