+++ tests/integration/AccountLinking.spec.ts
@@ -65,7 +65,7 @@ describe('When there is an on-chain DID', () => {
 
       const associateSenderTx = api.tx.didLookup.associateSender()
       const signedTx = await Did.authorizeTx(
-        did.uri,
+        did.id,
         associateSenderTx,
         didKey.getSignCallback(did),
         paymentAccount.address
@@ -92,12 +92,12 @@ describe('When there is an on-chain DID', () => {
       )
       const queryByAccount = Did.linkedInfoFromChain(encodedQueryByAccount)
       expect(queryByAccount.accounts).toStrictEqual([paymentAccount.address])
-      expect(queryByAccount.document.uri).toStrictEqual(did.uri)
+      expect(queryByAccount.document.id).toStrictEqual(did.id)
     }, 30_000)
     it('should be possible to associate the tx sender to a new DID', async () => {
       const associateSenderTx = api.tx.didLookup.associateSender()
       const signedTx = await Did.authorizeTx(
-        newDid.uri,
+        newDid.id,
         associateSenderTx,
         newDidKey.getSignCallback(newDid),
         paymentAccount.address
@@ -120,7 +120,7 @@ describe('When there is an on-chain DID', () => {
       )
       const queryByAccount = Did.linkedInfoFromChain(encodedQueryByAccount)
       expect(queryByAccount.accounts).toStrictEqual([paymentAccount.address])
-      expect(queryByAccount.document.uri).toStrictEqual(newDid.uri)
+      expect(queryByAccount.document.id).toStrictEqual(newDid.id)
     }, 30_000)
     it('should be possible for the sender to remove the link', async () => {
       const removeSenderTx = api.tx.didLookup.removeSenderAssociation()
@@ -174,11 +174,11 @@ describe('When there is an on-chain DID', () => {
         }
         const args = await Did.associateAccountToChainArgs(
           keypair.address,
-          did.uri,
+          did.id,
           async (payload) => keypair.sign(payload, { withType: false })
         )
         const signedTx = await Did.authorizeTx(
-          did.uri,
+          did.id,
           api.tx.didLookup.associateAccount(...args),
           didKey.getSignCallback(did),
           paymentAccount.address
@@ -204,7 +204,7 @@ describe('When there is an on-chain DID', () => {
         )
         const queryByAccount = Did.linkedInfoFromChain(encodedQueryByAccount)
         expect(queryByAccount.accounts).toStrictEqual([keypair.address])
-        expect(queryByAccount.document.uri).toStrictEqual(did.uri)
+        expect(queryByAccount.document.id).toStrictEqual(did.id)
       })
       it('should be possible to associate the account to a new DID while the sender pays the deposit', async () => {
         if (skip) {
@@ -212,11 +212,11 @@ describe('When there is an on-chain DID', () => {
         }
         const args = await Did.associateAccountToChainArgs(
           keypair.address,
-          newDid.uri,
+          newDid.id,
           async (payload) => keypair.sign(payload, { withType: false })
         )
         const signedTx = await Did.authorizeTx(
-          newDid.uri,
+          newDid.id,
           api.tx.didLookup.associateAccount(...args),
           newDidKey.getSignCallback(newDid),
           paymentAccount.address
@@ -239,7 +239,7 @@ describe('When there is an on-chain DID', () => {
         )
         const queryByAccount = Did.linkedInfoFromChain(encodedQueryByAccount)
         expect(queryByAccount.accounts).toStrictEqual([keypair.address])
-        expect(queryByAccount.document.uri).toStrictEqual(newDid.uri)
+        expect(queryByAccount.document.id).toStrictEqual(newDid.id)
       })
       it('should be possible for the DID to remove the link', async () => {
         if (skip) {
@@ -248,7 +248,7 @@ describe('When there is an on-chain DID', () => {
         const removeLinkTx =
           api.tx.didLookup.removeAccountAssociation(keypairChain)
         const signedTx = await Did.authorizeTx(
-          newDid.uri,
+          newDid.id,
           removeLinkTx,
           newDidKey.getSignCallback(newDid),
           paymentAccount.address
@@ -271,7 +271,7 @@ describe('When there is an on-chain DID', () => {
         )
         expect(encodedQueryByAccount.isNone).toBe(true)
         const encodedQueryByDid = await api.call.did.query(
-          Did.toChain(newDid.uri)
+          Did.toChain(newDid.id)
         )
         const queryByDid = Did.linkedInfoFromChain(encodedQueryByDid)
         expect(queryByDid.accounts).toStrictEqual([])
@@ -299,11 +299,11 @@ describe('When there is an on-chain DID', () => {
     it('should be possible to associate the account while the sender pays the deposit', async () => {
       const args = await Did.associateAccountToChainArgs(
         genericAccount.address,
-        did.uri,
+        did.id,
         async (payload) => genericAccount.sign(payload, { withType: true })
       )
       const signedTx = await Did.authorizeTx(
-        did.uri,
+        did.id,
         api.tx.didLookup.associateAccount(...args),
         didKey.getSignCallback(did),
         paymentAccount.address
@@ -330,13 +330,13 @@ describe('When there is an on-chain DID', () => {
       // Use generic substrate address prefix
       const queryByAccount = Did.linkedInfoFromChain(encodedQueryByAccount, 42)
       expect(queryByAccount.accounts).toStrictEqual([genericAccount.address])
-      expect(queryByAccount.document.uri).toStrictEqual(did.uri)
+      expect(queryByAccount.document.id).toStrictEqual(did.id)
     })
 
     it('should be possible to add a Web3 name for the linked DID and retrieve it starting from the linked account', async () => {
       const web3NameClaimTx = api.tx.web3Names.claim('test-name')
       const signedTx = await Did.authorizeTx(
-        did.uri,
+        did.id,
         web3NameClaimTx,
         didKey.getSignCallback(did),
         paymentAccount.address
@@ -346,13 +346,13 @@ describe('When there is an on-chain DID', () => {
       // Check that the Web3 name has been linked to the DID
       const encodedQueryByW3n = await api.call.did.queryByWeb3Name('test-name')
       const queryByW3n = Did.linkedInfoFromChain(encodedQueryByW3n)
-      expect(queryByW3n.document.uri).toStrictEqual(did.uri)
+      expect(queryByW3n.document.id).toStrictEqual(did.id)
       // Check that it is possible to retrieve the web3 name from the account linked to the DID
       const encodedQueryByAccount = await api.call.did.queryByAccount(
         Did.accountToChain(genericAccount.address)
       )
       const queryByAccount = Did.linkedInfoFromChain(encodedQueryByAccount)
-      expect(queryByAccount.web3Name).toStrictEqual('test-name')
+      expect(queryByAccount.document.alsoKnownAs).toStrictEqual(['test-name'])
     })
 
     it('should be possible for the sender to remove the link', async () => {
