+++ tests/integration/Attestation.spec.ts
@@ -77,7 +77,7 @@ describe('handling attestations that do not exist', () => {
   it('Attestation.getRevokeTx', async () => {
     const draft = api.tx.attestation.revoke(claimHash, null)
     const authorized = await Did.authorizeTx(
-      attester.uri,
+      attester.id,
       draft,
       attesterKey.getSignCallback(attester),
       tokenHolder.address
@@ -91,7 +91,7 @@ describe('handling attestations that do not exist', () => {
   it('Attestation.getRemoveTx', async () => {
     const draft = api.tx.attestation.remove(claimHash, null)
     const authorized = await Did.authorizeTx(
-      attester.uri,
+      attester.id,
       draft,
       attesterKey.getSignCallback(attester),
       tokenHolder.address
@@ -108,7 +108,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
     const ctypeExists = await isCtypeOnChain(driversLicenseCType)
     if (ctypeExists) return
     const tx = await Did.authorizeTx(
-      attester.uri,
+      attester.id,
       api.tx.ctype.add(CType.toChain(driversLicenseCType)),
       attesterKey.getSignCallback(attester),
       tokenHolder.address
@@ -121,7 +121,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
     const claim = Claim.fromCTypeAndClaimContents(
       driversLicenseCType,
       content,
-      claimer.uri
+      claimer.id
     )
     const credential = Credential.fromClaim(claim)
     const presentation = await Credential.createPresentation({
@@ -141,7 +141,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
     const claim = Claim.fromCTypeAndClaimContents(
       driversLicenseCType,
       content,
-      claimer.uri
+      claimer.id
     )
     const credential = Credential.fromClaim(claim)
     expect(() => Credential.verifyDataIntegrity(credential)).not.toThrow()
@@ -157,7 +157,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
 
     const attestation = Attestation.fromCredentialAndDid(
       presentation,
-      attester.uri
+      attester.id
     )
     const storeTx = api.tx.attestation.add(
       attestation.claimHash,
@@ -165,7 +165,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       null
     )
     const authorizedStoreTx = await Did.authorizeTx(
-      attester.uri,
+      attester.id,
       storeTx,
       attesterKey.getSignCallback(attester),
       tokenHolder.address
@@ -180,7 +180,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
 
     await expect(
       Credential.verifyPresentation(presentation)
-    ).resolves.toMatchObject({ attester: attester.uri, revoked: false })
+    ).resolves.toMatchObject({ attester: attester.id, revoked: false })
 
     // Claim the deposit back by submitting the reclaimDeposit extrinsic with the deposit payer's account.
     const reclaimTx = api.tx.attestation.reclaimDeposit(attestation.claimHash)
@@ -202,7 +202,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
     const claim = Claim.fromCTypeAndClaimContents(
       driversLicenseCType,
       content,
-      claimer.uri
+      claimer.id
     )
     const credential = Credential.fromClaim(claim)
     expect(() => Credential.verifyDataIntegrity(credential)).not.toThrow()
@@ -217,7 +217,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
 
     const attestation = Attestation.fromCredentialAndDid(
       presentation,
-      attester.uri
+      attester.id
     )
     const { keypair, getSignCallback } = makeSigningKeyTool()
 
@@ -227,7 +227,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       null
     )
     const authorizedStoreTx = await Did.authorizeTx(
-      attester.uri,
+      attester.id,
       storeTx,
       getSignCallback(attester),
       keypair.address
@@ -254,15 +254,11 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
     })
 
     const content = { name: 'Ralph', weight: 120 }
-    const claim = Claim.fromCTypeAndClaimContents(
-      badCtype,
-      content,
-      claimer.uri
-    )
+    const claim = Claim.fromCTypeAndClaimContents(badCtype, content, claimer.id)
     const credential = Credential.fromClaim(claim)
     const attestation = Attestation.fromCredentialAndDid(
       credential,
-      attester.uri
+      attester.id
     )
     const storeTx = api.tx.attestation.add(
       attestation.claimHash,
@@ -270,7 +266,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       null
     )
     const authorizedStoreTx = await Did.authorizeTx(
-      attester.uri,
+      attester.id,
       storeTx,
       attesterKey.getSignCallback(attester),
       tokenHolder.address
@@ -293,21 +289,21 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       const claim = Claim.fromCTypeAndClaimContents(
         driversLicenseCType,
         content,
-        claimer.uri
+        claimer.id
       )
       credential = Credential.fromClaim(claim)
       const presentation = await Credential.createPresentation({
         credential,
         signCallback: claimerKey.getSignCallback(claimer),
       })
-      attestation = Attestation.fromCredentialAndDid(credential, attester.uri)
+      attestation = Attestation.fromCredentialAndDid(credential, attester.id)
       const storeTx = api.tx.attestation.add(
         attestation.claimHash,
         attestation.cTypeHash,
         null
       )
       const authorizedStoreTx = await Did.authorizeTx(
-        attester.uri,
+        attester.id,
         storeTx,
         attesterKey.getSignCallback(attester),
         tokenHolder.address
@@ -330,7 +326,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
         null
       )
       const authorizedStoreTx = await Did.authorizeTx(
-        attester.uri,
+        attester.id,
         storeTx,
         attesterKey.getSignCallback(attester),
         tokenHolder.address
@@ -349,7 +345,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       const claim = Claim.fromCTypeAndClaimContents(
         driversLicenseCType,
         content,
-        claimer.uri
+        claimer.id
       )
       const fakeCredential = Credential.fromClaim(claim)
       await Credential.createPresentation({
@@ -365,7 +361,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
     it('should not be possible for the claimer to revoke an attestation', async () => {
       const revokeTx = api.tx.attestation.revoke(attestation.claimHash, null)
       const authorizedRevokeTx = await Did.authorizeTx(
-        claimer.uri,
+        claimer.id,
         revokeTx,
         claimerKey.getSignCallback(claimer),
         tokenHolder.address
@@ -395,7 +391,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
 
       const revokeTx = api.tx.attestation.revoke(attestation.claimHash, null)
       const authorizedRevokeTx = await Did.authorizeTx(
-        attester.uri,
+        attester.id,
         revokeTx,
         attesterKey.getSignCallback(attester),
         tokenHolder.address
@@ -411,13 +407,13 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
 
       await expect(
         Credential.verifyCredential(credential)
-      ).resolves.toMatchObject({ attester: attester.uri, revoked: true })
+      ).resolves.toMatchObject({ attester: attester.id, revoked: true })
     }, 40_000)
 
     it('should be possible for the deposit payer to remove an attestation', async () => {
       const removeTx = api.tx.attestation.remove(attestation.claimHash, null)
       const authorizedRemoveTx = await Did.authorizeTx(
-        attester.uri,
+        attester.id,
         removeTx,
         attesterKey.getSignCallback(attester),
         tokenHolder.address
@@ -446,7 +442,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
         CType.toChain(officialLicenseAuthorityCType)
       )
       const authorizedStoreTx = await Did.authorizeTx(
-        attester.uri,
+        attester.id,
         storeTx,
         attesterKey.getSignCallback(attester),
         tokenHolder.address
@@ -464,7 +460,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
           LicenseType: "Driver's License",
           LicenseSubtypes: 'sports cars, tanks',
         },
-        attester.uri
+        attester.id
       )
       const credential1 = Credential.fromClaim(licenseAuthorization)
       await Credential.createPresentation({
@@ -473,7 +469,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       })
       const licenseAuthorizationGranted = Attestation.fromCredentialAndDid(
         credential1,
-        anotherAttester.uri
+        anotherAttester.id
       )
       const storeTx = api.tx.attestation.add(
         licenseAuthorizationGranted.claimHash,
@@ -481,7 +477,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
         null
       )
       const authorizedStoreTx = await Did.authorizeTx(
-        anotherAttester.uri,
+        anotherAttester.id,
         storeTx,
         anotherAttesterKey.getSignCallback(anotherAttester),
         tokenHolder.address
@@ -492,7 +488,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       const iBelieveICanDrive = Claim.fromCTypeAndClaimContents(
         driversLicenseCType,
         { name: 'Dominic Toretto', age: 52 },
-        claimer.uri
+        claimer.id
       )
       const credential2 = Credential.fromClaim(iBelieveICanDrive, {
         legitimations: [credential1],
@@ -503,7 +499,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
       })
       const licenseGranted = Attestation.fromCredentialAndDid(
         credential2,
-        attester.uri
+        attester.id
       )
       const storeTx2 = api.tx.attestation.add(
         licenseGranted.claimHash,
@@ -511,7 +507,7 @@ describe('When there is an attester, claimer and ctype drivers license', () => {
         null
       )
       const authorizedStoreTx2 = await Did.authorizeTx(
-        attester.uri,
+        attester.id,
         storeTx2,
         attesterKey.getSignCallback(attester),
         tokenHolder.address
