+++ tests/integration/Did.spec.ts
@@ -6,27 +6,26 @@
  */
 
 import type { ApiPromise } from '@polkadot/api'
-import { BN } from '@polkadot/util'
-
-import { CType, DelegationNode, disconnect } from '@kiltprotocol/core'
-import * as Did from '@kiltprotocol/did'
-import {
+import type {
   DidDocument,
-  DidResolutionResult,
-  DidServiceEndpoint,
   KiltKeyringPair,
-  NewDidEncryptionKey,
-  NewDidVerificationKey,
-  NewLightDidVerificationKey,
-  Permission,
+  ResolutionResult,
+  Service,
   SignCallback,
+  VerificationMethod,
 } from '@kiltprotocol/types'
+
+import { BN } from '@polkadot/util'
+import { CType, DelegationNode, disconnect } from '@kiltprotocol/core'
+import { Permission } from '@kiltprotocol/types'
 import { UUID } from '@kiltprotocol/utils'
+import * as Did from '@kiltprotocol/did'
+
+import type { KeyTool } from '../testUtils/index.js'
 
 import {
   createFullDidFromSeed,
   createMinimalLightDidFromKeypair,
-  KeyTool,
   makeEncryptionKeyTool,
   makeSigningKeyTool,
 } from '../testUtils/index.js'
@@ -61,7 +60,7 @@ describe('write and didDeleteTx', () => {
 
   it('fails to create a new DID on chain with a different submitter than the one in the creation operation', async () => {
     const otherAccount = devBob
-    const tx = await Did.getStoreTx(
+    const tx = await Did.getStoreTxFromDidDocument(
       did,
       otherAccount.address,
       key.storeDidCallback
@@ -73,8 +72,15 @@ describe('write and didDeleteTx', () => {
   }, 60_000)
 
   it('writes a new DID record to chain', async () => {
+    const { publicKeyMultibase } = did.verificationMethod?.find(
+      (vm) => vm.id === did.authentication?.[0]
+    ) as VerificationMethod
+    const { keyType, publicKey: authPublicKey } =
+      Did.multibaseKeyToDidKey(publicKeyMultibase)
     const newDid = Did.createLightDidDocument({
-      authentication: did.authentication as [NewLightDidVerificationKey],
+      authentication: [{ publicKey: authPublicKey, type: keyType }] as [
+        Did.NewDidVerificationKey
+      ],
       service: [
         {
           id: '#test-id-1',
@@ -89,7 +95,7 @@ describe('write and didDeleteTx', () => {
       ],
     })
 
-    const tx = await Did.getStoreTx(
+    const tx = await Did.getStoreTxFromDidDocument(
       newDid,
       paymentAccount.address,
       key.storeDidCallback
@@ -97,19 +103,12 @@ describe('write and didDeleteTx', () => {
 
     await submitTx(tx, paymentAccount)
 
-    const fullDidUri = Did.getFullDidUri(newDid.uri)
+    const fullDidUri = Did.getFullDidUri(newDid.id)
     const fullDidLinkedInfo = await api.call.did.query(Did.toChain(fullDidUri))
     const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
 
-    expect(fullDid).toMatchObject(<DidDocument>{
-      uri: fullDidUri,
-      authentication: [
-        expect.objectContaining({
-          // We cannot match the ID of the key because it will be defined by the blockchain while saving
-          publicKey: newDid.authentication[0].publicKey,
-          type: 'sr25519',
-        }),
-      ],
+    expect(fullDid).toMatchObject(<Partial<DidDocument>>{
+      id: fullDidUri,
       service: [
         {
           id: '#test-id-1',
@@ -122,13 +121,26 @@ describe('write and didDeleteTx', () => {
           type: ['test-type-2'],
         },
       ],
+      verificationMethod: [
+        expect.objectContaining(<Partial<VerificationMethod>>{
+          controller: fullDidUri,
+          type: 'MultiKey',
+          // We cannot match the ID of the key because it will be defined by the blockchain while saving
+          publicKeyMultibase: Did.keypairToMultibaseKey({
+            type: 'sr25519',
+            publicKey: authPublicKey,
+          }),
+        }),
+      ],
     })
+    expect(fullDid.authentication).toHaveLength(1)
+    expect(fullDid.keyAgreement).toBe(undefined)
+    expect(fullDid.assertionMethod).toBe(undefined)
+    expect(fullDid.capabilityDelegation).toBe(undefined)
   }, 60_000)
 
   it('should return no results for empty accounts', async () => {
-    const emptyDid = Did.getFullDidUriFromKey(
-      makeSigningKeyTool().authentication[0]
-    )
+    const emptyDid = Did.getFullDidUri(makeSigningKeyTool().keypair.address)
 
     const encodedDid = Did.toChain(emptyDid)
     expect((await api.call.did.query(encodedDid)).isSome).toBe(false)
@@ -137,7 +149,7 @@ describe('write and didDeleteTx', () => {
   it('fails to delete the DID using a different submitter than the one specified in the DID operation or using a services count that is too low', async () => {
     // We verify that the DID to delete is on chain.
     const fullDidLinkedInfo = await api.call.did.query(
-      Did.toChain(Did.getFullDidUri(did.uri))
+      Did.toChain(Did.getFullDidUri(did.id))
     )
     const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
     expect(fullDid).not.toBeNull()
@@ -148,7 +160,7 @@ describe('write and didDeleteTx', () => {
     let call = api.tx.did.delete(new BN(10))
 
     let submittable = await Did.authorizeTx(
-      fullDid.uri,
+      fullDid.id,
       call,
       signCallback,
       // Use a different account than the submitter one
@@ -164,7 +176,7 @@ describe('write and didDeleteTx', () => {
     call = api.tx.did.delete(new BN(1))
 
     submittable = await Did.authorizeTx(
-      fullDid.uri,
+      fullDid.id,
       call,
       signCallback,
       paymentAccount.address
@@ -182,12 +194,12 @@ describe('write and didDeleteTx', () => {
   it('deletes DID from previous step', async () => {
     // We verify that the DID to delete is on chain.
     const fullDidLinkedInfo = await api.call.did.query(
-      Did.toChain(Did.getFullDidUri(did.uri))
+      Did.toChain(Did.getFullDidUri(did.id))
     )
     const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
     expect(fullDid).not.toBeNull()
 
-    const encodedDid = Did.toChain(fullDid.uri)
+    const encodedDid = Did.toChain(fullDid.id)
     const linkedInfo = Did.linkedInfoFromChain(
       await api.call.did.query(encodedDid)
     )
@@ -195,7 +207,7 @@ describe('write and didDeleteTx', () => {
     const call = api.tx.did.delete(storedEndpointsCount)
 
     const submittable = await Did.authorizeTx(
-      fullDid.uri,
+      fullDid.id,
       call,
       signCallback,
       paymentAccount.address
@@ -217,7 +229,7 @@ it('creates and updates DID, and then reclaims the deposit back', async () => {
   const { keypair, getSignCallback, storeDidCallback } = makeSigningKeyTool()
   const newDid = await createMinimalLightDidFromKeypair(keypair)
 
-  const tx = await Did.getStoreTx(
+  const tx = await Did.getStoreTxFromDidDocument(
     newDid,
     paymentAccount.address,
     storeDidCallback
@@ -227,7 +239,7 @@ it('creates and updates DID, and then reclaims the deposit back', async () => {
 
   // This will better be handled once we have the UpdateBuilder class, which encapsulates all the logic.
   let fullDidLinkedInfo = await api.call.did.query(
-    Did.toChain(Did.getFullDidUri(newDid.uri))
+    Did.toChain(Did.getFullDidUri(newDid.id))
   )
   let { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
 
@@ -237,7 +249,7 @@ it('creates and updates DID, and then reclaims the deposit back', async () => {
     Did.publicKeyToChain(newKey.authentication[0])
   )
   const tx2 = await Did.authorizeTx(
-    fullDid.uri,
+    fullDid.id,
     updateAuthenticationKeyCall,
     getSignCallback(fullDid),
     paymentAccount.address
@@ -247,12 +259,12 @@ it('creates and updates DID, and then reclaims the deposit back', async () => {
   // Authentication key changed, so did must be updated.
   // Also this will better be handled once we have the UpdateBuilder class, which encapsulates all the logic.
   fullDidLinkedInfo = await api.call.did.query(
-    Did.toChain(Did.getFullDidUri(newDid.uri))
+    Did.toChain(Did.getFullDidUri(newDid.id))
   )
   fullDid = Did.linkedInfoFromChain(fullDidLinkedInfo).document
 
   // Add a new service endpoint
-  const newEndpoint: DidServiceEndpoint = {
+  const newEndpoint: Did.NewService = {
     id: '#new-endpoint',
     type: ['new-type'],
     serviceEndpoint: ['x:new-url'],
@@ -262,27 +274,27 @@ it('creates and updates DID, and then reclaims the deposit back', async () => {
   )
 
   const tx3 = await Did.authorizeTx(
-    fullDid.uri,
+    fullDid.id,
     updateEndpointCall,
     newKey.getSignCallback(fullDid),
     paymentAccount.address
   )
   await submitTx(tx3, paymentAccount)
 
-  const encodedDid = Did.toChain(fullDid.uri)
+  const encodedDid = Did.toChain(fullDid.id)
   const linkedInfo = Did.linkedInfoFromChain(
     await api.call.did.query(encodedDid)
   )
-  expect(Did.getService(linkedInfo.document, newEndpoint.id)).toStrictEqual(
-    newEndpoint
-  )
+  expect(
+    linkedInfo.document.service?.find((s) => s.id === newEndpoint.id)
+  ).toStrictEqual(newEndpoint)
 
   // Delete the added service endpoint
   const removeEndpointCall = api.tx.did.removeServiceEndpoint(
-    Did.resourceIdToChain(newEndpoint.id)
+    Did.fragmentIdToChain(newEndpoint.id)
   )
   const tx4 = await Did.authorizeTx(
-    fullDid.uri,
+    fullDid.id,
     removeEndpointCall,
     newKey.getSignCallback(fullDid),
     paymentAccount.address
@@ -293,7 +305,9 @@ it('creates and updates DID, and then reclaims the deposit back', async () => {
   const linkedInfo2 = Did.linkedInfoFromChain(
     await api.call.did.query(encodedDid)
   )
-  expect(Did.getService(linkedInfo2.document, newEndpoint.id)).toBe(undefined)
+  expect(
+    linkedInfo2.document.service?.find((s) => s.id === newEndpoint.id)
+  ).toBe(undefined)
 
   // Claim the deposit back
   const storedEndpointsCount = linkedInfo2.document.service?.length ?? 0
@@ -317,14 +331,14 @@ describe('DID migration', () => {
       keyAgreement,
     })
 
-    const storeTx = await Did.getStoreTx(
+    const storeTx = await Did.getStoreTxFromDidDocument(
       lightDid,
       paymentAccount.address,
       storeDidCallback
     )
 
     await submitTx(storeTx, paymentAccount)
-    const migratedFullDidUri = Did.getFullDidUri(lightDid.uri)
+    const migratedFullDidUri = Did.getFullDidUri(lightDid.id)
     const migratedFullDidLinkedInfo = await api.call.did.query(
       Did.toChain(migratedFullDidUri)
     )
@@ -332,32 +346,38 @@ describe('DID migration', () => {
       migratedFullDidLinkedInfo
     )
 
-    expect(migratedFullDid).toMatchObject(<DidDocument>{
-      uri: migratedFullDidUri,
-      authentication: [
-        expect.objectContaining({
-          publicKey: lightDid.authentication[0].publicKey,
-          type: 'ed25519',
+    expect(migratedFullDid).toMatchObject(<Partial<DidDocument>>{
+      id: migratedFullDidUri,
+      verificationMethod: [
+        expect.objectContaining(<Partial<VerificationMethod>>{
+          controller: migratedFullDidUri,
+          type: 'MultiKey',
+          // We cannot match the ID of the key because it will be defined by the blockchain while saving
+          publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
         }),
-      ],
-      keyAgreement: [
-        expect.objectContaining({
-          publicKey: lightDid.keyAgreement?.[0].publicKey,
-          type: 'x25519',
+        expect.objectContaining(<Partial<VerificationMethod>>{
+          controller: migratedFullDidUri,
+          type: 'MultiKey',
+          // We cannot match the ID of the key because it will be defined by the blockchain while saving
+          publicKeyMultibase: Did.keypairToMultibaseKey(keyAgreement[0]),
         }),
       ],
     })
+    expect(migratedFullDid.authentication).toHaveLength(1)
+    expect(migratedFullDid.keyAgreement).toHaveLength(1)
+    expect(migratedFullDid.assertionMethod).toBe(undefined)
+    expect(migratedFullDid.capabilityDelegation).toBe(undefined)
 
     expect(
-      (await api.call.did.query(Did.toChain(migratedFullDid.uri))).isSome
+      (await api.call.did.query(Did.toChain(migratedFullDid.id))).isSome
     ).toBe(true)
 
-    const { metadata } = (await Did.resolve(
-      lightDid.uri
-    )) as DidResolutionResult
+    const { didDocumentMetadata } = (await Did.resolve(
+      lightDid.id
+    )) as ResolutionResult
 
-    expect(metadata.canonicalId).toStrictEqual(migratedFullDid.uri)
-    expect(metadata.deactivated).toBe(false)
+    expect(didDocumentMetadata.canonicalId).toStrictEqual(migratedFullDid.id)
+    expect(didDocumentMetadata.deactivated).toBe(undefined)
   })
 
   it('migrates light DID with sr25519 auth key', async () => {
@@ -366,14 +386,14 @@ describe('DID migration', () => {
       authentication,
     })
 
-    const storeTx = await Did.getStoreTx(
+    const storeTx = await Did.getStoreTxFromDidDocument(
       lightDid,
       paymentAccount.address,
       storeDidCallback
     )
 
     await submitTx(storeTx, paymentAccount)
-    const migratedFullDidUri = Did.getFullDidUri(lightDid.uri)
+    const migratedFullDidUri = Did.getFullDidUri(lightDid.id)
     const migratedFullDidLinkedInfo = await api.call.did.query(
       Did.toChain(migratedFullDidUri)
     )
@@ -381,26 +401,32 @@ describe('DID migration', () => {
       migratedFullDidLinkedInfo
     )
 
-    expect(migratedFullDid).toMatchObject(<DidDocument>{
-      uri: migratedFullDidUri,
-      authentication: [
-        expect.objectContaining({
-          publicKey: lightDid.authentication[0].publicKey,
-          type: 'sr25519',
+    expect(migratedFullDid).toMatchObject(<Partial<DidDocument>>{
+      id: migratedFullDidUri,
+      verificationMethod: [
+        expect.objectContaining(<Partial<VerificationMethod>>{
+          controller: migratedFullDidUri,
+          type: 'MultiKey',
+          // We cannot match the ID of the key because it will be defined by the blockchain while saving
+          publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
         }),
       ],
     })
+    expect(migratedFullDid.authentication).toHaveLength(1)
+    expect(migratedFullDid.keyAgreement).toBe(undefined)
+    expect(migratedFullDid.assertionMethod).toBe(undefined)
+    expect(migratedFullDid.capabilityDelegation).toBe(undefined)
 
     expect(
-      (await api.call.did.query(Did.toChain(migratedFullDid.uri))).isSome
+      (await api.call.did.query(Did.toChain(migratedFullDid.id))).isSome
     ).toBe(true)
 
-    const { metadata } = (await Did.resolve(
-      lightDid.uri
-    )) as DidResolutionResult
+    const { didDocumentMetadata } = (await Did.resolve(
+      lightDid.id
+    )) as ResolutionResult
 
-    expect(metadata.canonicalId).toStrictEqual(migratedFullDid.uri)
-    expect(metadata.deactivated).toBe(false)
+    expect(didDocumentMetadata.canonicalId).toStrictEqual(migratedFullDid.id)
+    expect(didDocumentMetadata.deactivated).toBe(undefined)
   })
 
   it('migrates light DID with ed25519 auth key, encryption key, and service endpoints', async () => {
@@ -408,7 +434,7 @@ describe('DID migration', () => {
     const { keyAgreement } = makeEncryptionKeyTool(
       '0xcccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc'
     )
-    const service: DidServiceEndpoint[] = [
+    const service: Did.NewService[] = [
       {
         id: '#id-1',
         type: ['type-1'],
@@ -421,14 +447,14 @@ describe('DID migration', () => {
       service,
     })
 
-    const storeTx = await Did.getStoreTx(
+    const storeTx = await Did.getStoreTxFromDidDocument(
       lightDid,
       paymentAccount.address,
       storeDidCallback
     )
 
     await submitTx(storeTx, paymentAccount)
-    const migratedFullDidUri = Did.getFullDidUri(lightDid.uri)
+    const migratedFullDidUri = Did.getFullDidUri(lightDid.id)
     const migratedFullDidLinkedInfo = await api.call.did.query(
       Did.toChain(migratedFullDidUri)
     )
@@ -436,18 +462,20 @@ describe('DID migration', () => {
       migratedFullDidLinkedInfo
     )
 
-    expect(migratedFullDid).toMatchObject(<DidDocument>{
-      uri: migratedFullDidUri,
-      authentication: [
-        expect.objectContaining({
-          publicKey: lightDid.authentication[0].publicKey,
-          type: 'ed25519',
+    expect(migratedFullDid).toMatchObject(<Partial<DidDocument>>{
+      id: migratedFullDidUri,
+      verificationMethod: [
+        expect.objectContaining(<Partial<VerificationMethod>>{
+          controller: migratedFullDidUri,
+          type: 'MultiKey',
+          // We cannot match the ID of the key because it will be defined by the blockchain while saving
+          publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
         }),
-      ],
-      keyAgreement: [
-        expect.objectContaining({
-          publicKey: lightDid.keyAgreement?.[0].publicKey,
-          type: 'x25519',
+        expect.objectContaining(<Partial<VerificationMethod>>{
+          controller: migratedFullDidUri,
+          type: 'MultiKey',
+          // We cannot match the ID of the key because it will be defined by the blockchain while saving
+          publicKeyMultibase: Did.keypairToMultibaseKey(keyAgreement[0]),
         }),
       ],
       service: [
@@ -458,16 +486,20 @@ describe('DID migration', () => {
         },
       ],
     })
+    expect(migratedFullDid.authentication).toHaveLength(1)
+    expect(migratedFullDid.keyAgreement).toHaveLength(1)
+    expect(migratedFullDid.assertionMethod).toBe(undefined)
+    expect(migratedFullDid.capabilityDelegation).toBe(undefined)
 
-    const encodedDid = Did.toChain(migratedFullDid.uri)
+    const encodedDid = Did.toChain(migratedFullDid.id)
     expect((await api.call.did.query(encodedDid)).isSome).toBe(true)
 
-    const { metadata } = (await Did.resolve(
-      lightDid.uri
-    )) as DidResolutionResult
+    const { didDocumentMetadata } = (await Did.resolve(
+      lightDid.id
+    )) as ResolutionResult
 
-    expect(metadata.canonicalId).toStrictEqual(migratedFullDid.uri)
-    expect(metadata.deactivated).toBe(false)
+    expect(didDocumentMetadata.canonicalId).toStrictEqual(migratedFullDid.id)
+    expect(didDocumentMetadata.deactivated).toBe(undefined)
 
     // Remove and claim the deposit back
     const linkedInfo = Did.linkedInfoFromChain(
@@ -492,7 +524,7 @@ describe('DID authorization', () => {
     makeSigningKeyTool('ed25519')
 
   beforeAll(async () => {
-    const createTx = await Did.getStoreTx(
+    const createTx = await Did.getStoreTxFromInput(
       {
         authentication,
         assertionMethod: authentication,
@@ -503,7 +535,11 @@ describe('DID authorization', () => {
     )
     await submitTx(createTx, paymentAccount)
     const didLinkedInfo = await api.call.did.query(
-      Did.toChain(Did.getFullDidUriFromKey(authentication[0]))
+      Did.toChain(
+        Did.getFullDidUriFromVerificationMethod({
+          publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+        })
+      )
     )
     did = Did.linkedInfoFromChain(didLinkedInfo).document
   }, 60_000)
@@ -512,7 +548,7 @@ describe('DID authorization', () => {
     const cType = CType.fromProperties(UUID.generate(), {})
     const call = api.tx.ctype.add(CType.toChain(cType))
     const tx = await Did.authorizeTx(
-      did.uri,
+      did.id,
       call,
       getSignCallback(did),
       paymentAccount.address
@@ -524,12 +560,12 @@ describe('DID authorization', () => {
 
   it('no longer authorizes ctype creation after DID deletion', async () => {
     const linkedInfo = Did.linkedInfoFromChain(
-      await api.call.did.query(Did.toChain(did.uri))
+      await api.call.did.query(Did.toChain(did.id))
     )
     const storedEndpointsCount = linkedInfo.document.service?.length ?? 0
     const deleteCall = api.tx.did.delete(storedEndpointsCount)
     const tx = await Did.authorizeTx(
-      did.uri,
+      did.id,
       deleteCall,
       getSignCallback(did),
       paymentAccount.address
@@ -539,7 +575,7 @@ describe('DID authorization', () => {
     const cType = CType.fromProperties(UUID.generate(), {})
     const call = api.tx.ctype.add(CType.toChain(cType))
     const tx2 = await Did.authorizeTx(
-      did.uri,
+      did.id,
       call,
       getSignCallback(did),
       paymentAccount.address
@@ -556,8 +592,8 @@ describe('DID authorization', () => {
 describe('DID management batching', () => {
   describe('FullDidCreationBuilder', () => {
     it('Build a complete full DID', async () => {
-      const { keypair, storeDidCallback, authentication } = makeSigningKeyTool()
-      const extrinsic = await Did.getStoreTx(
+      const { storeDidCallback, authentication } = makeSigningKeyTool()
+      const extrinsic = await Did.getStoreTxFromInput(
         {
           authentication,
           assertionMethod: [
@@ -609,44 +645,71 @@ describe('DID management batching', () => {
       )
       await submitTx(extrinsic, paymentAccount)
       const fullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(Did.getFullDidUriFromKey(authentication[0]))
+        Did.toChain(
+          Did.getFullDidUriFromVerificationMethod({
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+          })
+        )
       )
       const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
 
       expect(fullDid).not.toBeNull()
-      expect(fullDid).toMatchObject({
-        authentication: [
+      expect(fullDid.verificationMethod).toEqual<Partial<VerificationMethod[]>>(
+        expect.arrayContaining([
           expect.objectContaining({
-            publicKey: keypair.publicKey,
-            type: 'sr25519',
+            // Authentication
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
           }),
-        ],
-        assertionMethod: [
+          // Assertion method
           expect.objectContaining({
-            publicKey: new Uint8Array(32).fill(1),
-            type: 'sr25519',
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              type: 'sr25519',
+              publicKey: new Uint8Array(32).fill(1),
+            }),
           }),
-        ],
-        capabilityDelegation: [
+          // Capability delegation
           expect.objectContaining({
-            publicKey: new Uint8Array(33).fill(1),
-            type: 'ecdsa',
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              type: 'ecdsa',
+              publicKey: new Uint8Array(33).fill(1),
+            }),
           }),
-        ],
-        keyAgreement: [
+          // Key agreement 1
           expect.objectContaining({
-            publicKey: new Uint8Array(32).fill(3),
-            type: 'x25519',
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              type: 'x25519',
+              publicKey: new Uint8Array(32).fill(1),
+            }),
           }),
+          // Key agreement 2
           expect.objectContaining({
-            publicKey: new Uint8Array(32).fill(2),
-            type: 'x25519',
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              type: 'x25519',
+              publicKey: new Uint8Array(32).fill(2),
+            }),
           }),
+          // Key agreement 3
           expect.objectContaining({
-            publicKey: new Uint8Array(32).fill(1),
-            type: 'x25519',
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              type: 'x25519',
+              publicKey: new Uint8Array(32).fill(3),
+            }),
           }),
-        ],
+        ])
+      )
+      expect(fullDid).toMatchObject(<Partial<DidDocument>>{
         service: [
           {
             id: '#id-3',
@@ -665,16 +728,20 @@ describe('DID management batching', () => {
           },
         ],
       })
+      expect(fullDid.authentication).toHaveLength(1)
+      expect(fullDid.assertionMethod).toHaveLength(1)
+      expect(fullDid.capabilityDelegation).toHaveLength(1)
+      expect(fullDid.keyAgreement).toHaveLength(3)
     })
 
     it('Build a minimal full DID with an Ecdsa key', async () => {
       const { keypair, storeDidCallback } = makeSigningKeyTool('ecdsa')
-      const didAuthKey: NewDidVerificationKey = {
+      const didAuthKey: Did.NewDidVerificationKey = {
         publicKey: keypair.publicKey,
         type: 'ecdsa',
       }
 
-      const extrinsic = await Did.getStoreTx(
+      const extrinsic = await Did.getStoreTxFromInput(
         { authentication: [didAuthKey] },
         paymentAccount.address,
         storeDidCallback
@@ -682,17 +749,29 @@ describe('DID management batching', () => {
       await submitTx(extrinsic, paymentAccount)
 
       const fullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(Did.getFullDidUriFromKey(didAuthKey))
+        Did.toChain(
+          Did.getFullDidUriFromVerificationMethod({
+            publicKeyMultibase: Did.keypairToMultibaseKey(didAuthKey),
+          })
+        )
       )
       const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
 
       expect(fullDid).not.toBeNull()
-      expect(fullDid?.authentication).toMatchObject<NewDidVerificationKey[]>([
-        {
-          publicKey: keypair.publicKey,
-          type: 'ecdsa',
-        },
-      ])
+      expect(fullDid).toMatchObject(<Partial<DidDocument>>{
+        verificationMethod: [
+          // Authentication
+          expect.objectContaining(<Partial<VerificationMethod>>{
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey(didAuthKey),
+          }),
+        ],
+      })
+      expect(fullDid.authentication).toHaveLength(1)
+      expect(fullDid.assertionMethod).toBe(undefined)
+      expect(fullDid.capabilityDelegation).toBe(undefined)
+      expect(fullDid.keyAgreement).toBe(undefined)
     })
   })
 
@@ -701,7 +780,7 @@ describe('DID management batching', () => {
       const { keypair, getSignCallback, storeDidCallback, authentication } =
         makeSigningKeyTool()
 
-      const createTx = await Did.getStoreTx(
+      const createTx = await Did.getStoreTxFromInput(
         {
           authentication,
           keyAgreement: [
@@ -745,7 +824,11 @@ describe('DID management batching', () => {
       await submitTx(createTx, paymentAccount)
 
       const initialFullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(Did.getFullDidUriFromKey(authentication[0]))
+        Did.toChain(
+          Did.getFullDidUriFromVerificationMethod({
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+          })
+        )
       )
       const { document: initialFullDid } = Did.linkedInfoFromChain(
         initialFullDidLinkedInfo
@@ -756,13 +839,21 @@ describe('DID management batching', () => {
 
       const extrinsic = await Did.authorizeBatch({
         batchFunction: api.tx.utility.batchAll,
-        did: initialFullDid.uri,
+        did: initialFullDid.id,
         extrinsics: [
           api.tx.did.removeKeyAgreementKey(
-            Did.resourceIdToChain(encryptionKeys[0].id)
+            Did.fragmentIdToChain(
+              initialFullDid.verificationMethod!.find(
+                (vm) => vm.id === encryptionKeys[0]
+              )!.id
+            )
           ),
           api.tx.did.removeKeyAgreementKey(
-            Did.resourceIdToChain(encryptionKeys[1].id)
+            Did.fragmentIdToChain(
+              initialFullDid.verificationMethod!.find(
+                (vm) => vm.id === encryptionKeys[1]
+              )!.id
+            )
           ),
           api.tx.did.removeAttestationKey(),
           api.tx.did.removeDelegationKey(),
@@ -775,7 +866,7 @@ describe('DID management batching', () => {
       await submitTx(extrinsic, paymentAccount)
 
       const finalFullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(initialFullDid.uri)
+        Did.toChain(initialFullDid.id)
       )
       const { document: finalFullDid } = Did.linkedInfoFromChain(
         finalFullDidLinkedInfo
@@ -783,17 +874,23 @@ describe('DID management batching', () => {
 
       expect(finalFullDid).not.toBeNull()
 
-      expect(
-        finalFullDid.authentication[0]
-      ).toMatchObject<NewDidVerificationKey>({
-        publicKey: keypair.publicKey,
-        type: 'sr25519',
+      expect(finalFullDid).toMatchObject(<Partial<DidDocument>>{
+        verificationMethod: [
+          // Authentication
+          expect.objectContaining(<Partial<VerificationMethod>>{
+            controller: finalFullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              publicKey: keypair.publicKey,
+              type: 'sr25519',
+            }),
+          }),
+        ],
       })
-
-      expect(finalFullDid.keyAgreement).toBeUndefined()
-      expect(finalFullDid.assertionMethod).toBeUndefined()
-      expect(finalFullDid.capabilityDelegation).toBeUndefined()
-      expect(finalFullDid.service).toBeUndefined()
+      expect(finalFullDid.authentication).toHaveLength(1)
+      expect(finalFullDid.assertionMethod).toBe(undefined)
+      expect(finalFullDid.capabilityDelegation).toBe(undefined)
+      expect(finalFullDid.keyAgreement).toBe(undefined)
     }, 40_000)
 
     it('Correctly handles rotation of the authentication key', async () => {
@@ -801,9 +898,9 @@ describe('DID management batching', () => {
         makeSigningKeyTool()
       const {
         authentication: [newAuthKey],
-      } = makeSigningKeyTool('ed25519')
+      } = TestUtils.makeSigningKeyTool('ed25519')
 
-      const createTx = await Did.getStoreTx(
+      const createTx = await Did.getStoreTxFromInput(
         { authentication },
         paymentAccount.address,
         storeDidCallback
@@ -811,7 +908,11 @@ describe('DID management batching', () => {
       await submitTx(createTx, paymentAccount)
 
       const initialFullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(Did.getFullDidUriFromKey(authentication[0]))
+        Did.toChain(
+          Did.getFullDidUriFromVerificationMethod({
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+          })
+        )
       )
       const { document: initialFullDid } = Did.linkedInfoFromChain(
         initialFullDidLinkedInfo
@@ -819,7 +920,7 @@ describe('DID management batching', () => {
 
       const extrinsic = await Did.authorizeBatch({
         batchFunction: api.tx.utility.batchAll,
-        did: initialFullDid.uri,
+        did: initialFullDid.id,
         extrinsics: [
           api.tx.did.addServiceEndpoint(
             Did.serviceToChain({
@@ -844,29 +945,49 @@ describe('DID management batching', () => {
       await submitTx(extrinsic, paymentAccount)
 
       const finalFullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(initialFullDid.uri)
+        Did.toChain(initialFullDid.id)
       )
       const { document: finalFullDid } = Did.linkedInfoFromChain(
         finalFullDidLinkedInfo
       )
 
       expect(finalFullDid).not.toBeNull()
-
-      expect(finalFullDid.authentication[0]).toMatchObject({
-        publicKey: newAuthKey.publicKey,
-        type: newAuthKey.type,
+      expect(finalFullDid).toMatchObject(<Partial<DidDocument>>{
+        verificationMethod: [
+          // Authentication
+          expect.objectContaining(<Partial<VerificationMethod>>{
+            controller: finalFullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey({
+              publicKey: newAuthKey.publicKey,
+              type: 'ed25519',
+            }),
+          }),
+        ],
+        service: [
+          {
+            id: '#id-1',
+            type: ['type-1'],
+            serviceEndpoint: ['x:url-1'],
+          },
+          {
+            id: '#id-2',
+            type: ['type-2'],
+            serviceEndpoint: ['x:url-2'],
+          },
+        ],
       })
 
+      expect(finalFullDid.authentication).toHaveLength(1)
       expect(finalFullDid.keyAgreement).toBeUndefined()
       expect(finalFullDid.assertionMethod).toBeUndefined()
       expect(finalFullDid.capabilityDelegation).toBeUndefined()
-      expect(finalFullDid.service).toHaveLength(2)
     }, 40_000)
 
     it('simple `batch` succeeds despite failures of some extrinsics', async () => {
       const { authentication, getSignCallback, storeDidCallback } =
         makeSigningKeyTool()
-      const tx = await Did.getStoreTx(
+      const tx = await Did.getStoreTxFromInput(
         {
           authentication,
           service: [
@@ -883,7 +1004,11 @@ describe('DID management batching', () => {
       // Create the full DID with a service endpoint
       await submitTx(tx, paymentAccount)
       const fullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(Did.getFullDidUriFromKey(authentication[0]))
+        Did.toChain(
+          Did.getFullDidUriFromVerificationMethod({
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+          })
+        )
       )
       const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
 
@@ -892,7 +1017,7 @@ describe('DID management batching', () => {
       // Try to set a new attestation key and a duplicate service endpoint
       const updateTx = await Did.authorizeBatch({
         batchFunction: api.tx.utility.batch,
-        did: fullDid.uri,
+        did: fullDid.id,
         extrinsics: [
           api.tx.did.setAttestationKey(Did.publicKeyToChain(authentication[0])),
           api.tx.did.addServiceEndpoint(
@@ -910,28 +1035,44 @@ describe('DID management batching', () => {
       await submitTx(updateTx, paymentAccount)
 
       const updatedFullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(fullDid.uri)
+        Did.toChain(fullDid.id)
       )
       const { document: updatedFullDid } = Did.linkedInfoFromChain(
         updatedFullDidLinkedInfo
       )
 
-      // .setAttestationKey() extrinsic went through in the batch
-      expect(updatedFullDid.assertionMethod?.[0]).toBeDefined()
-      // The service endpoint will match the one manually added, and not the one set in the batch
-      expect(
-        Did.getService(updatedFullDid, '#id-1')
-      ).toStrictEqual<DidServiceEndpoint>({
-        id: '#id-1',
-        type: ['type-1'],
-        serviceEndpoint: ['x:url-1'],
+      expect(updatedFullDid).toMatchObject<Partial<DidDocument>>({
+        verificationMethod: [
+          expect.objectContaining({
+            // Authentication and assertionMethod
+            controller: fullDid.id,
+            type: 'MultiKey',
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+          }),
+        ],
+        // Old service maintained
+        service: [
+          {
+            id: '#id-1',
+            type: ['type-1'],
+            serviceEndpoint: ['x:url-1'],
+          },
+        ],
       })
+
+      expect(updatedFullDid.authentication).toHaveLength(1)
+      expect(updatedFullDid.keyAgreement).toBeUndefined()
+      // .setAttestationKey() extrinsic went through in the batch
+      expect(updatedFullDid.assertionMethod).toStrictEqual(
+        updatedFullDid.authentication
+      )
+      expect(updatedFullDid.capabilityDelegation).toBeUndefined()
     }, 60_000)
 
     it('batchAll fails if any extrinsics fails', async () => {
       const { authentication, getSignCallback, storeDidCallback } =
         makeSigningKeyTool()
-      const createTx = await Did.getStoreTx(
+      const createTx = await Did.getStoreTxFromInput(
         {
           authentication,
           service: [
@@ -947,7 +1088,11 @@ describe('DID management batching', () => {
       )
       await submitTx(createTx, paymentAccount)
       const fullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(Did.getFullDidUriFromKey(authentication[0]))
+        Did.toChain(
+          Did.getFullDidUriFromVerificationMethod({
+            publicKeyMultibase: Did.keypairToMultibaseKey(authentication[0]),
+          })
+        )
       )
       const { document: fullDid } = Did.linkedInfoFromChain(fullDidLinkedInfo)
 
@@ -956,7 +1101,7 @@ describe('DID management batching', () => {
       // Use batchAll to set a new attestation key and a duplicate service endpoint
       const updateTx = await Did.authorizeBatch({
         batchFunction: api.tx.utility.batchAll,
-        did: fullDid.uri,
+        did: fullDid.id,
         extrinsics: [
           api.tx.did.setAttestationKey(Did.publicKeyToChain(authentication[0])),
           api.tx.did.addServiceEndpoint(
@@ -978,7 +1123,7 @@ describe('DID management batching', () => {
       })
 
       const updatedFullDidLinkedInfo = await api.call.did.query(
-        Did.toChain(fullDid.uri)
+        Did.toChain(fullDid.id)
       )
       const { document: updatedFullDid } = Did.linkedInfoFromChain(
         updatedFullDidLinkedInfo
@@ -987,8 +1132,8 @@ describe('DID management batching', () => {
       expect(updatedFullDid.assertionMethod).toBeUndefined()
       // The service endpoint will match the one manually added, and not the one set in the builder.
       expect(
-        Did.getService(updatedFullDid, '#id-1')
-      ).toStrictEqual<DidServiceEndpoint>({
+        updatedFullDid.service?.find((s) => s.id === '#id-1')
+      ).toStrictEqual<Service>({
         id: '#id-1',
         type: ['type-1'],
         serviceEndpoint: ['x:url-1'],
@@ -1010,15 +1155,15 @@ describe('DID extrinsics batching', () => {
     const cType = CType.fromProperties(UUID.generate(), {})
     const ctypeStoreTx = api.tx.ctype.add(CType.toChain(cType))
     const rootNode = DelegationNode.newRoot({
-      account: fullDid.uri,
+      account: fullDid.id,
       permissions: [Permission.DELEGATE],
       cTypeHash: CType.idToHash(cType.$id),
     })
     const delegationStoreTx = await rootNode.getStoreTx()
-    const delegationRevocationTx = await rootNode.getRevokeTx(fullDid.uri)
+    const delegationRevocationTx = await rootNode.getRevokeTx(fullDid.id)
     const tx = await Did.authorizeBatch({
       batchFunction: api.tx.utility.batch,
-      did: fullDid.uri,
+      did: fullDid.id,
       extrinsics: [
         ctypeStoreTx,
         // Will fail since the delegation cannot be revoked before it is added
@@ -1040,15 +1185,15 @@ describe('DID extrinsics batching', () => {
     const cType = CType.fromProperties(UUID.generate(), {})
     const ctypeStoreTx = api.tx.ctype.add(CType.toChain(cType))
     const rootNode = DelegationNode.newRoot({
-      account: fullDid.uri,
+      account: fullDid.id,
       permissions: [Permission.DELEGATE],
       cTypeHash: CType.idToHash(cType.$id),
     })
     const delegationStoreTx = await rootNode.getStoreTx()
-    const delegationRevocationTx = await rootNode.getRevokeTx(fullDid.uri)
+    const delegationRevocationTx = await rootNode.getRevokeTx(fullDid.id)
     const tx = await Did.authorizeBatch({
       batchFunction: api.tx.utility.batchAll,
-      did: fullDid.uri,
+      did: fullDid.id,
       extrinsics: [
         ctypeStoreTx,
         // Will fail since the delegation cannot be revoked before it is added
@@ -1072,7 +1217,7 @@ describe('DID extrinsics batching', () => {
   it('can batch extrinsics for the same required key type', async () => {
     const web3NameClaimTx = api.tx.web3Names.claim('test-1')
     const authorizedTx = await Did.authorizeTx(
-      fullDid.uri,
+      fullDid.id,
       web3NameClaimTx,
       key.getSignCallback(fullDid),
       paymentAccount.address
@@ -1083,7 +1228,7 @@ describe('DID extrinsics batching', () => {
     const web3Name2ClaimExt = api.tx.web3Names.claim('test-2')
     const tx = await Did.authorizeBatch({
       batchFunction: api.tx.utility.batch,
-      did: fullDid.uri,
+      did: fullDid.id,
       extrinsics: [web3Name1ReleaseExt, web3Name2ClaimExt],
       sign: key.getSignCallback(fullDid),
       submitter: paymentAccount.address,
@@ -1095,8 +1240,8 @@ describe('DID extrinsics batching', () => {
     expect(encoded1.isSome).toBe(false)
     // Test for correct creation of second web3 name
     const encoded2 = await api.call.did.queryByWeb3Name('test-2')
-    expect(Did.linkedInfoFromChain(encoded2).document.uri).toStrictEqual(
-      fullDid.uri
+    expect(Did.linkedInfoFromChain(encoded2).document.id).toStrictEqual(
+      fullDid.id
     )
   }, 30_000)
 
@@ -1108,7 +1253,7 @@ describe('DID extrinsics batching', () => {
     const ctype1Creation = api.tx.ctype.add(CType.toChain(ctype1))
     // Delegation key
     const rootNode = DelegationNode.newRoot({
-      account: fullDid.uri,
+      account: fullDid.id,
       permissions: [Permission.DELEGATE],
       cTypeHash: CType.idToHash(ctype1.$id),
     })
@@ -1120,11 +1265,11 @@ describe('DID extrinsics batching', () => {
     const ctype2 = CType.fromProperties(UUID.generate(), {})
     const ctype2Creation = api.tx.ctype.add(CType.toChain(ctype2))
     // Delegation key
-    const delegationHierarchyRemoval = await rootNode.getRevokeTx(fullDid.uri)
+    const delegationHierarchyRemoval = await rootNode.getRevokeTx(fullDid.id)
 
     const batchedExtrinsics = await Did.authorizeBatch({
       batchFunction: api.tx.utility.batchAll,
-      did: fullDid.uri,
+      did: fullDid.id,
       extrinsics: [
         web3NameReleaseExt,
         ctype1Creation,
@@ -1144,9 +1289,9 @@ describe('DID extrinsics batching', () => {
     expect(encoded.isSome).toBe(false)
 
     const {
-      document: { uri },
+      document: { id },
     } = Did.linkedInfoFromChain(await api.call.did.queryByWeb3Name('test-2'))
-    expect(uri).toStrictEqual(fullDid.uri)
+    expect(id).toStrictEqual(fullDid.id)
 
     // Test correct use of attestation keys
     await expect(CType.verifyStored(ctype1)).resolves.not.toThrow()
@@ -1159,7 +1304,7 @@ describe('DID extrinsics batching', () => {
 })
 
 describe('Runtime constraints', () => {
-  let testAuthKey: NewDidVerificationKey
+  let testAuthKey: Did.NewDidVerificationKey
   const { keypair, storeDidCallback } = makeSigningKeyTool('ed25519')
 
   beforeAll(async () => {
@@ -1172,12 +1317,12 @@ describe('Runtime constraints', () => {
     it('should not be possible to create a DID with too many encryption keys', async () => {
       // Maximum is 10
       const newKeyAgreementKeys = Array(10).map(
-        (_, index): NewDidEncryptionKey => ({
+        (_, index): Did.NewDidEncryptionKey => ({
           publicKey: Uint8Array.from(new Array(32).fill(index)),
           type: 'x25519',
         })
       )
-      await Did.getStoreTx(
+      await Did.getStoreTxFromInput(
         {
           authentication: [testAuthKey],
           keyAgreement: newKeyAgreementKeys,
@@ -1191,7 +1336,7 @@ describe('Runtime constraints', () => {
         type: 'x25519',
       })
       await expect(
-        Did.getStoreTx(
+        Did.getStoreTxFromInput(
           {
             authentication: [testAuthKey],
             keyAgreement: newKeyAgreementKeys,
@@ -1208,13 +1353,13 @@ describe('Runtime constraints', () => {
     it('should not be possible to create a DID with too many service endpoints', async () => {
       // Maximum is 25
       const newServiceEndpoints = Array(25).map(
-        (_, index): DidServiceEndpoint => ({
+        (_, index): Did.NewService => ({
           id: `#service-${index}`,
           type: [`type-${index}`],
           serviceEndpoint: [`x:url-${index}`],
         })
       )
-      await Did.getStoreTx(
+      await Did.getStoreTxFromInput(
         {
           authentication: [testAuthKey],
           service: newServiceEndpoints,
@@ -1229,7 +1374,7 @@ describe('Runtime constraints', () => {
         serviceEndpoint: ['x:url-100'],
       })
       await expect(
-        Did.getStoreTx(
+        Did.getStoreTxFromInput(
           {
             authentication: [testAuthKey],
             service: newServiceEndpoints,
